Main: fix warning (size_t assigned to ULONG)
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@11093 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/MachineImpl.cpp
@@ -8900,7 +8900,7 @@ STDMETHODIMP SessionMachine::PullGuestProperties (ComSafeArrayOut(BSTR, aNames),
     AssertReturn(!ComSafeArrayOutIsNull (aTimestamps), E_POINTER);
     AssertReturn(!ComSafeArrayOutIsNull (aFlags), E_POINTER);
 
-    unsigned cEntries = mHWData->mGuestProperties.size();
+    size_t cEntries = mHWData->mGuestProperties.size();
     com::SafeArray <BSTR> names(cEntries);
     com::SafeArray <BSTR> values(cEntries);
     com::SafeArray <ULONG64> timestamps(cEntries);
